refactor(main): separate DesktopBackendBackground into its own file
